refactored some more
